feat(#682): improved reusage for mutations
