+++ packages/core/src/governance/client/hooks/useDeployFundsMutations.web3.rpc.ts
@@ -5,11 +5,12 @@ import { Config, writeContract } from '@wagmi/core';
 import { getProposalFromLogs } from '../web3';
 import useSWR from 'swr';
 import { publicClient } from '@core/common/web3/public-client';
-import { encodeFunctionData, erc20Abi, getContract, parseUnits } from 'viem';
+import { encodeFunctionData, erc20Abi, parseUnits } from 'viem';
 import {
   daoProposalsImplementationAbi,
   daoProposalsImplementationAddress,
 } from '@core/generated';
+import { getTokenDecimals } from './get-token-decimals';
 
 interface CreateDeployFundsInput {
   spaceId: number;
@@ -18,16 +19,6 @@ interface CreateDeployFundsInput {
     token: string;
   }[];
   recipient: string;
-}
-
-async function getTokenDecimals(tokenAddress: string): Promise<number> {
-  const contract = getContract({
-    address: tokenAddress as `0x${string}`,
-    abi: erc20Abi,
-    client: publicClient,
-  });
-
-  return await contract.read.decimals();
 }
 
 export const useDeployFundsMutationsWeb3Rpc = (config?: Config) => {
